==> 02_network_and_practical/src/01_downloading_the_dataset.ipynb <==
## **Introduction to ML for NLP [Network + Practical]**

### **Downloading the Dataset**

All information on the chosen dataset can be read [here](https://huggingface.co/datasets/multi_eurlex), on the official HuggingFace website.

#### **Libraries**

We import the necessary libraries for the notebook.

# general
import pandas as pd
from tqdm import tqdm

# library with multi_eurlex data
from datasets import load_dataset

print("> Libraries Imported")

#### **Import the dataset**

There are 23 official EU languages available, but we only choose 5:

| **_Language_** | **_Total EU Speakers (%)_** |
|:--------------:|:---------------------------:|
|        English | 51                          |
|         German | 32                          |
|        Italian | 16                          |
|         Polish | 9                           |
|        Swedish | 3                           |

*Source: [HuggingFace](https://huggingface.co/datasets/multi_eurlex)*

**English**

Let's start with english.

dataset_en = load_dataset('multi_eurlex', 'en')

We transform it in a pandas dataframe, to make it easier to explore and work with.

# explore the structure
dataset_en

We create a custom function to convert the data.

def from_hf_dataset_to_dataframe(dataset, language):

    # list of available keys
    keys = ["train","test","validation"]

    # create placeholder
    celex_id = []
    text = []
    labels = []

    # iterate over each key
    for key in keys:

        # iterate over each id, text and labels for key
        custom_desc = "Converting '" + key + "' rows in a dataframe"
        for item in tqdm(dataset[key], total=len(dataset[key]), desc=custom_desc):

            celex_id.append(item["celex_id"])
            text.append(item["text"])
            labels.append(item["labels"])

    # create the dataframe, starting from the 3 lists
    final_df = pd.DataFrame(
        list(zip(celex_id, text, labels)),
        columns =['celex_id', 'text_' + language, 'labels_' + language]
        )

    # finally, return the df
    return final_df

dataframe_en = from_hf_dataset_to_dataframe(dataset_en, language="en")
dataframe_en

**German**

We now do the same for German.

First, we download the data.

dataset_de = load_dataset('multi_eurlex', 'de')

Then we convert it to a dataframe using our custom function.

dataframe_de = from_hf_dataset_to_dataframe(dataset_de, language="de")
dataframe_de

The number of observations is the same. 

However, we perform an inner join on `celex_id` in order to have the same sentences in both languages (just to be sure).

dataframe_complete = pd.merge(
    dataframe_en,
    dataframe_de,
    on='celex_id',
    how="inner"
    )

dataframe_complete

As expected, the number of observations did not change.

Note: we keep the labels in all languages, although they should obviously match, in order to check (once all joins have been made) that there are no inconsistencies in the labels of the final dataset. 

**Italian, Polish, Swedish**

We now apply the df conversion and inner joins to the texts in the remaining languages.

dataset_it = load_dataset('multi_eurlex', 'it')
dataset_pl = load_dataset('multi_eurlex', 'pl')
dataset_sv = load_dataset('multi_eurlex', 'sv')

dataframe_it = from_hf_dataset_to_dataframe(dataset_it, language="it")
dataframe_pl = from_hf_dataset_to_dataframe(dataset_pl, language="pl")
dataframe_sv = from_hf_dataset_to_dataframe(dataset_sv, language="sv")

# italian
dataframe_complete = pd.merge(
    dataframe_complete,
    dataframe_it,
    on='celex_id',
    how="inner"
    )

# polish
dataframe_complete = pd.merge(
    dataframe_complete,
    dataframe_pl,
    on='celex_id',
    how="inner"
    )

# swedish
dataframe_complete = pd.merge(
    dataframe_complete,
    dataframe_sv,
    on='celex_id',
    how="inner"
    )

dataframe_complete

With less popular languages, such as Italian, Polish and Swedish, not all texts are available. 

After the inner joins we are left with around 33k observations, which are more than enough to continue our classification.  

**Are there any inconsistencies?**

Before saving the final dataset, we make sure that there are no inconsistencies in the labels.

inconsistent_rows = []

for id,row in dataframe_complete.iterrows():

    if row["labels_en"] != row["labels_de"] or \
       row["labels_en"] != row["labels_it"] or \
       row["labels_en"] != row["labels_pl"] or \
       row["labels_en"] != row["labels_sv"]:

        inconsistent_rows.append(id)

# show inconsistent rows, if any
inconsistent_rows

There are no inconsistencies, we can now save the complete dataframe to disk, as a .csv, for future use.

# remove redundant columns (we just need one 'labels' col)
dataframe_complete["labels"] = dataframe_complete["labels_en"]
dataframe_complete_clean = dataframe_complete[["celex_id","labels","text_en","text_de","text_it","text_pl","text_sv"]]

# show df
dataframe_complete_clean

Finally, save to csv.

dataframe_complete_clean.to_csv("data/0_multi_eurlex_custom.csv", index=False)
print("> All done!")
