==> 02_network_and_practical/src/02_data_cleaning.ipynb <==
## **Introduction to ML for NLP [Network + Practical]**

### **Cleaning the Dataset**

After obtaining the dataset with the texts to be classified in 5 different languages, we have to preprocess it. 

#### **Libraries**

We import the necessary libraries for the notebook.

# general
import pandas as pd
from tqdm import tqdm
tqdm.pandas()

# string manipulation
import re
import unidecode

print("> Libraries Imported")

#### **Import the dataset**

We can simply read our custom csv. 

dataframe_complete = pd.read_csv("data/0_multi_eurlex_custom.csv")
dataframe_complete

#### **Clean the texts**

First, let's inspect some texts.

dataframe_complete["text_en"][0]

There do not seem to be any particular problems, except for the special character *'\n'*, which we can simply replace with a space.

We can also remove the punctuation, as well as replace all numbers with a keyword.

**Note**

Usually, stopwords are removed and stemming or lemmatisation is applied. However, this would require knowing the list of stopwords for each language, along with specific dictionaries for stemming or lemmatization.

For now, we decide not to apply these procedures. By not applying them to any language, we will be able to compare the performance of the models without having advantaged any language (in terms of quality of the stopwords removed, etc.).

It is time to create our custom cleaning function.

NUMBER_RULE = re.compile("\d+")
ONLY_LETTERS = re.compile('[^a-zA-Z _]+')

def clean_text(text):

    text = text.replace("\n"," ")                   # substitute \n with a space
    #text = re.sub(NUMBER_RULE, " _number_ ", text)  # identify numbers in the text
    text = unidecode.unidecode(text)                # remove accents and sort (they may create errors in the future architectures)
    text = text.lower()                             # lowercase text
    text = ONLY_LETTERS.sub(' ', text)              # remove all that it is not text
    text = re.sub(' +', ' ', text)                  # remove multiple spaces, if any

    # returned pre-processed text
    return text

Let's try it out on the first text.

clean_text(dataframe_complete["text_en"][0])

It works as expected! We can now apply it to all the texts in the dataset.

# specify columns with texts
text_cols = ["text_en", "text_de", "text_it", "text_pl", "text_sv"]

# iterate over the text columns
for text_col in text_cols:
    print("\n > Cleaning column '" + text_col + "'")
    dataframe_complete[text_col] = dataframe_complete[text_col].progress_apply(clean_text)

dataframe_complete

#### **From multi-label to multi-class**

The dataset is designed for multi-label classification. However, as explained in the [paper](https://arxiv.org/pdf/2109.00904.pdf) on the dataset, the labels are *multi-granular*. 

This means that we can keep only the first label and convert the task into a multi-class classification.

We create a function that can extract the first value from the list. 

def clean_labels(labels):

    # first, we convert the string back to a list
    labels = list(labels[1:-1].split(", "))

    # then, we extract the first element only and we return it (as int)
    return int(labels[0])

We can then apply it to the whole dataset.

dataframe_complete["labels"] = dataframe_complete["labels"].progress_apply(clean_labels)

The paper says that there are 21 classes in the first level. 

We check that everything matches, showing the different available classes. 

set(dataframe_complete["labels"])

They are 21 indeed.

#### **Drop duplicates texts**

Are there duplicates in the texts?

print("> Total Observations:", dataframe_complete.shape[0])
dataframe_complete.drop_duplicates()
print("> Total Observations after Duplicates Drop:", dataframe_complete.shape[0])

Nice, there are no duplicates.

#### **Show cleaned dataset**

We now show the final dataset, on which we will train the architectures.

dataframe_complete

The cleaning procedure is time consuming (especially for the texts).

We save the dataset so that we can use it later.

dataframe_complete.to_csv("data/1_multi_eurlex_clean.csv", index=False)
print("> All done!")
