==> 02_network_and_practical/src/03_data_exploration.ipynb <==
## **Introduction to ML for NLP [Network + Practical]**

### **Exploring the Dataset**

After cleaning the dataset, we explore it in order to understand what data we are dealing with.

#### **Libraries**

We import the necessary libraries for the notebook.

# general
import pandas as pd
import numpy as np
from tqdm import tqdm
tqdm.pandas()

# dataviz
import plotly.express as px

print("> Libraries Imported")

#### **Import the dataset**

We can simply read our custom (and cleaned) csv. 

dataframe = pd.read_csv("data/1_multi_eurlex_clean.csv")
dataframe.head()

#### **Data Exploration**

**1. How many occurrences for class?**

Let's see wether the classes are balanced or not.

First, a simple function to plot the number of texts for each class.

def create_obs_per_classes_barplot(df, width, height, title):
    """
    Input:
        > df            dataset [tickets vendita]
        > width         of plot
        > height        of plot
        > title         of plot
    Output:
        > prints and returns the plot object
    """

    # data manipulation for plot
    temp_df = pd.DataFrame(df.groupby("labels")["text_en"].count()).reset_index()

    # convert labels to string (so that they are treated as categoricals)
    temp_df.labels = temp_df.labels.astype(str)

    # create plot
    fig = px.bar(
        temp_df, 
        x="labels", 
        y="text_en", 
        color="labels",
        title=title,
        width=width, height=height, 
        text_auto=True,
            labels={
                "text": "Number of Observations",
                "labels":"Label (first level)"
            },
        )
    fig.update_layout(xaxis_tickangle=-45, showlegend=False)
    fig.update_traces(textangle=0, textposition="outside", cliponaxis=False)
    fig.show()

    return fig

Now we can plot!

# create plot
barplot = create_obs_per_classes_barplot(
    df = dataframe,
    width = 1400, 
    height = 600, 
    title = "Number of Texts for each Label<br><sup>Total Classes: 21<sup>"
    )

The dataset is very unbalanced: a high number of classes have less than 500 observations, while others exceed 2000 (in some cases by a lot, i.e. with class 3).

As interesting as it is to test how architectures behave with unbalanced data, classes such as 14 would only introduce noise, given the low number of observations. We therefore decide to update the dataset, keeping only the classes with at least 500 observations.

**Note after the first training tests**: the texts are very, very long (1000-1200 words). As a consequence, the training times and epochs needed to arrive at satisfactory performance are high as well. 

In order to make the training faster, we select only 3 classes: *2*, *3* and *18* (the most frequent). For each class, we then randomly select 2000 observations each to avoid working with unbalanced classes. 

# 1. groupby texts
temp_df = pd.DataFrame(dataframe.groupby("labels")["text_en"].count()).reset_index()

# 2. obtain the list of labels with at least 500 obs
classes_to_keep = list(temp_df.loc[temp_df["text_en"] >= 2646]["labels"])

classes_to_keep

# filter the dataset
reduced_dataframe = dataframe.loc[dataframe["labels"].isin(classes_to_keep)]
reduced_dataframe

We now keep 2000 obs for each class.

dataframe_class2 = reduced_dataframe.loc[reduced_dataframe["labels"] == 2].sample(n=2000, random_state=42)
dataframe_class3 = reduced_dataframe.loc[reduced_dataframe["labels"] == 3].sample(n=2000, random_state=42) 
dataframe_class18 = reduced_dataframe.loc[reduced_dataframe["labels"] == 18].sample(n=2000, random_state=42) 

reduced_dataframe = pd.concat([dataframe_class2, dataframe_class3, dataframe_class18])
reduced_dataframe

Let's see if everything worked correctly with a plot :)

# create plot
barplot = create_obs_per_classes_barplot(
    df = reduced_dataframe,
    width = 1400, 
    height = 600, 
    title = "Number of Texts for each Label<br><sup>Total Classes: 3<sup>"
    )

#### **Re-map classes**

In order not to have errors during training, it is necessary to remap the label names (so that they go from 0 to 11).

As we already did before, we can simply apply a custom function.

# Reduced to 4 classes (> 2000 occurrences, class 3 excluded)
def map_old_labels_to_new_labels(label):

    if label == 2:
        return 0
    elif label == 3:
        return 1
    elif label == 18:
        return 2
    else:
        return label

# turn off pandas warning
pd.options.mode.chained_assignment = None

# apply the function
reduced_dataframe["labels_new"] = reduced_dataframe["labels"].progress_apply(map_old_labels_to_new_labels)

# explore results
set(reduced_dataframe["labels_new"])

**2. How many words in each text?**

text_cols = ["text_en", "text_de", "text_it", "text_pl", "text_sv"]

for col in text_cols:

    # obtain the number of words in each text
    reduced_dataframe[col + "_len"] = reduced_dataframe[col].apply(lambda x: len(x.split()))
    
    # calculate its mean
    temp_mean = np.mean(reduced_dataframe[col + "_len"])

    # print res
    print(f"> Average number of words in '{col[-2:]}' texts: {round(temp_mean)}")

The average number of words in the texts is around 900-1100. We will need this information in the tokenization and embedding phase.

#### **Save the changes**

Finally, we can save this final version of the dataset and train some networks!

# reorder columns
reduced_dataframe = reduced_dataframe[["celex_id", "labels", "labels_new", "text_en", "text_de", "text_it", "text_pl", "text_sv"]]

reduced_dataframe.to_csv("data/2_multi_eurlex_reduced.csv", index=False)
print("> All done!")
