==> 02_network_and_practical/src/04_preprocessing.ipynb <==
## **Introduction to ML for NLP [Network + Practical]**

### **Preprocessing**

It is now time to preprocess the texts.

#### **Libraries**

We import the necessary libraries for the notebook.

# general
import pandas as pd
import numpy as np
from tqdm import tqdm
tqdm.pandas()

# dataviz
import plotly.express as px         # create interactive plots

# preprocessing
import spacy                        # tokenization
from collections import Counter     # count words occurrences

print("> Libraries Imported")

#### **Import the dataset**

We read our custom (and cleaned) csv. 

dataframe = pd.read_csv("data/2_multi_eurlex_reduced.csv")
dataframe

#### **Preprocessing**

**Part 1: Tokenization**

This essentially involves dividing a sentence, paragraph or entire text document into smaller units, such as individual words or terms.

We create a custom function capable of tokenizing the texts in different languages.

# load the tokenizers (one for each language)

TOKENIZER_EN = spacy.load('en_core_web_sm')
TOKENIZER_DE = spacy.load('nl_core_news_sm')
TOKENIZER_IT = spacy.load('it_core_news_sm')
TOKENIZER_PL = spacy.load('pl_core_news_sm')
TOKENIZER_SV = spacy.load('sv_core_news_sm')

# create custom function

def tokenize(text, language):

    # use the required tokenizer (based on input language)
    if language=="en":
        tokenizer = TOKENIZER_EN
    elif language=="de":
        tokenizer = TOKENIZER_DE
    elif language=="it":
        tokenizer = TOKENIZER_IT
    elif language=="pl":
        tokenizer = TOKENIZER_PL
    elif language=="sv":
        tokenizer = TOKENIZER_SV
    else:
        error_str = "Error: language '" + language + "' not available. Please choose between 'en', 'de', 'it', 'pl', 'sv'."
        print(error_str)

    text = text.replace("_","")
    return [token.text for token in tokenizer.tokenizer(text)]

Here we show a small example of tokenization for the first text (in each language). 

test_text_en = dataframe["text_en"][0]
test_text_de = dataframe["text_de"][0]
test_text_it = dataframe["text_it"][0]
test_text_pl = dataframe["text_pl"][0]
test_text_sv = dataframe["text_sv"][0]


print("Sample of 'en' text:", test_text_en[0:120])
print("Sample of 'de' text:", test_text_de[0:120])
print("Sample of 'it' text:", test_text_it[0:120])
print("Sample of 'pl' text:", test_text_pl[0:120])
print("Sample of 'sv' text:", test_text_sv[0:120])

print("Example of tokenization in 'en':", tokenize(text=test_text_en, language="en")[0:10])
print("Example of tokenization in 'de':", tokenize(text=test_text_de, language="de")[0:10])
print("Example of tokenization in 'it':", tokenize(text=test_text_it, language="it")[0:10])
print("Example of tokenization in 'pl':", tokenize(text=test_text_pl, language="pl")[0:10])
print("Example of tokenization in 'sv':", tokenize(text=test_text_sv, language="sv")[0:10])

**Part 2: Count Words Occurrences**

We now count the number of occurrences of each word (now a token) in the corpuses.

# instantiate a Counter
COUNTS_EN = Counter()
COUNTS_DE = Counter()
COUNTS_IT = Counter()
COUNTS_PL = Counter()
COUNTS_SV = Counter()

for index, row in tqdm(dataframe.iterrows(), total=dataframe.shape[0], desc="> Counting words in texts"):
    COUNTS_EN.update(tokenize(row['text_en'], language="en"))
    COUNTS_DE.update(tokenize(row['text_de'], language="de"))
    COUNTS_IT.update(tokenize(row['text_it'], language="it"))
    COUNTS_PL.update(tokenize(row['text_pl'], language="pl"))
    COUNTS_SV.update(tokenize(row['text_sv'], language="sv"))

# show first 10 words of COUNTS_EN
i = 0
for key,value in COUNTS_EN.items():
    if i < 10:
        print(f"{key}: {value}")
    i+=1

*How many words in the dictionary of each language?*

print("Total Words in 'en' texts:", len(COUNTS_EN.keys()))
print("Total Words in 'de' texts:", len(COUNTS_DE.keys()))
print("Total Words in 'it' texts:", len(COUNTS_IT.keys()))
print("Total Words in 'pl' texts:", len(COUNTS_PL.keys()))
print("Total Words in 'sv' texts:", len(COUNTS_SV.keys()))

It is very interesting to note how the number of words used in each language changes significantly to describe the same laws. 

Considering English and Swedish, for example, we can see how the number of German words used exceeds the number of English words by a factor of 3.

The number of words is very high, let us try to remove the uncommon words.

COUNTS_EN = {key:val for key, val in COUNTS_EN.items() if val >= 100}
COUNTS_DE = {key:val for key, val in COUNTS_DE.items() if val >= 100}
COUNTS_IT = {key:val for key, val in COUNTS_IT.items() if val >= 100}
COUNTS_PL = {key:val for key, val in COUNTS_PL.items() if val >= 100}
COUNTS_SV = {key:val for key, val in COUNTS_SV.items() if val >= 100}

Let's observe the changes.

print("Total Words in 'en' texts:", len(COUNTS_EN.keys()))
print("Total Words in 'de' texts:", len(COUNTS_DE.keys()))
print("Total Words in 'it' texts:", len(COUNTS_IT.keys()))
print("Total Words in 'pl' texts:", len(COUNTS_PL.keys()))
print("Total Words in 'sv' texts:", len(COUNTS_SV.keys()))

Starting from the generated dictionaries, we can now create the 5 vocabularies.

# prepare placeholders
# - one for index mapping
# - one with the list of all the words

EN_WORDS_TO_INDEX = {"":0, "UNK":1}
EN_WORDS = ["", "UNK"]

DE_WORDS_TO_INDEX = {"":0, "UNK":1}
DE_WORDS = ["", "UNK"]

IT_WORDS_TO_INDEX = {"":0, "UNK":1}
IT_WORDS = ["", "UNK"]

PL_WORDS_TO_INDEX = {"":0, "UNK":1}
PL_WORDS = ["", "UNK"]

SV_WORDS_TO_INDEX = {"":0, "UNK":1}
SV_WORDS = ["", "UNK"]


# iterate over each counter and populate the placeholders

for word in COUNTS_EN:
    EN_WORDS_TO_INDEX[word] = len(EN_WORDS)
    EN_WORDS.append(word)

for word in COUNTS_DE:
    DE_WORDS_TO_INDEX[word] = len(DE_WORDS)
    DE_WORDS.append(word)

for word in COUNTS_IT:
    IT_WORDS_TO_INDEX[word] = len(IT_WORDS)
    IT_WORDS.append(word)

for word in COUNTS_PL:
    PL_WORDS_TO_INDEX[word] = len(PL_WORDS)
    PL_WORDS.append(word)

for word in COUNTS_SV:
    SV_WORDS_TO_INDEX[word] = len(SV_WORDS)
    SV_WORDS.append(word)

print("> Dictionaries Generated")

print("Total Words in 'en' texts:", len(EN_WORDS))
print("Total Words in 'de' texts:", len(DE_WORDS))
print("Total Words in 'it' texts:", len(IT_WORDS))
print("Total Words in 'pl' texts:", len(PL_WORDS))
print("Total Words in 'sv' texts:", len(SV_WORDS))

**Part 3: Encoding**

We have everything we need to encode our texts!

Note: we choose that the maximum length of each review should be 1000 words, as the average length of texts is around that value.

# create our custom encoding function

def encode_text(text, word_to_index, language, N=1000):
    
    # 1. tokenize the text
    tokenized = tokenize(text, language)

    # 2. encode the text
    encoded = np.zeros(N, dtype=int)
    enc1 = np.array([word_to_index.get(word, word_to_index["UNK"]) for word in tokenized])
    length = min(N, len(enc1))
    encoded[:length] = enc1[:length]

    # finally, return encoded text
    return encoded, length

# test the function on the first english text

test_encoded = encode_text(
    text=dataframe["text_en"][0], 
    word_to_index=EN_WORDS_TO_INDEX, 
    language="en", 
    N=1000
    )

test_encoded

# finally, we can apply the function to the whole df, encoding all texts

dataframe['text_en_enc'] = dataframe['text_en'].progress_apply(lambda x: np.array(encode_text(x, EN_WORDS_TO_INDEX, language="en"), dtype=object))
dataframe['text_de_enc'] = dataframe['text_de'].progress_apply(lambda x: np.array(encode_text(x, DE_WORDS_TO_INDEX, language="de"), dtype=object))
dataframe['text_it_enc'] = dataframe['text_it'].progress_apply(lambda x: np.array(encode_text(x, IT_WORDS_TO_INDEX, language="it"), dtype=object))
dataframe['text_pl_enc'] = dataframe['text_pl'].progress_apply(lambda x: np.array(encode_text(x, PL_WORDS_TO_INDEX, language="pl"), dtype=object))
dataframe['text_sv_enc'] = dataframe['text_sv'].progress_apply(lambda x: np.array(encode_text(x, SV_WORDS_TO_INDEX, language="sv"), dtype=object))

# show df
dataframe

#### **Train, Test and Validation splits**

We split the dataset into:
- training set
- validation set
- test set

from sklearn.model_selection import train_test_split

# SPLIT 1: train [0.8] and test [0.2]

# execute the split (randomly, stratified fashion)
train_indexes, test_indexes, _, _ = train_test_split(
    dataframe["text_en_enc"],
    dataframe["labels_new"],
    test_size=0.2,
    random_state=42
)

# create placeholder
dataframe["set"] = "not specified"

# set value based on stratified split
dataframe.loc[train_indexes.index, "set"] = "train"
dataframe.loc[test_indexes.index, "set"] = "test"


# SPLIT 2: we split train again in order to obtain the validation set

train_indexes, val_indexes, _, _ = train_test_split(
    dataframe.loc[dataframe["set"] == "train"]["text_en_enc"],
    dataframe.loc[dataframe["set"] == "train"]["labels_new"],
    test_size=0.2,
    random_state=42
)

# update dataset with validation info
dataframe.loc[val_indexes.index, "set"] = "validation"

*How many observations for each set?*

Counter(dataframe['set'])

#### **Save preprocessed texts**

We save the data in pickle format in order to preserve the column types.

dataframe.to_pickle("data/3_multi_eurlex_encoded.pkl")
print("> All done!")
