==> 02_network_and_practical/src/06.A_CNN_fixed_length.ipynb <==
## **Introduction to ML for NLP [Network + Practical]**

### **CNN**

It is now time to train our Convolutional Neural Nework.

#### **Libraries**

We import the necessary libraries for the notebook.

# general
import pandas as pd
from tqdm import tqdm
tqdm.pandas()

# pytorch
import torch

# custom imports
from utility.models_pytorch import PytorchModel
from utility.dataviz import plot_model_fit_loss, plot_classes_accuracy

print("> Libraries Imported")

#### **Setup**

- We set the device to *cuda*
- We import the dataset

# Device configuration
device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')

print("> Device:", device)

dataframe = pd.read_pickle("data/3_multi_eurlex_encoded.pkl")
dataframe.head(3)

#### **LSTM**

**Instantiate a Pytorch Model**

We use our custom class PytorchModel to train a LSTM.

COUNTS_EN = 3506
COUNTS_DE = 4216
COUNTS_IT = 4180
COUNTS_PL = 5255
COUNTS_SV = 4010

CNN_MODEL = PytorchModel(

    # set model and text language
    model_type      = "CNN_fixed",
    dataset         = dataframe,
    language        = "en",

    # set device, bacth size and epochs
    device          = device,
    batch_size      = 64,
    epochs          = 50,

    # set general hyperparameters
    learning_rate   = 0.001,

    # set specific hyperparameters
    vocab_size      = COUNTS_EN,
    embedding_dim   = 1024,
    out_channels    = 1,
    kernel_size     = 5,
    stride          = 1,
    padding         = 2,
    dropout_p       = 0.1,
)

CNN_MODEL = PytorchModel(

    # set model and text language
    model_type      = "CNN_fixed",
    dataset         = dataframe,
    language        = "en",

    # set device, bacth size and epochs
    device          = device,
    batch_size      = 64,
    epochs          = 50,

    # set general hyperparameters
    learning_rate   = 0.001,

    # set specific hyperparameters
    vocab_size      = COUNTS_EN,
    embedding_dim   = 1024,
    out_channels    = 1,
    kernel_size     = 5,
    stride          = 1,
    padding         = 2,
    dropout_p       = 0.1,
)

**Train the model**

We can now train the model.

The method will evaluate the performance of the model for each epoch.

global_res_df, classes_res_df = CNN_MODEL.train_model()

global_res_df

classes_res_df

**Visualize the training results**

We plot the training and validation loss, as well as the mean validation accuracy for each class.

plot_model_fit_loss(
    train_loss=global_res_df['training_loss'],
    val_loss=global_res_df['validation_loss'],
    subtitle="Models Details: " + CNN_MODEL.MODEL_DESCRIPTION
)

plot_classes_accuracy(
    classes_res_df, 
    subtitle="Models Details: " + CNN_MODEL.MODEL_DESCRIPTION
    )
