==> 02_network_and_practical/src/07.B_LinearSVC_grid_search.ipynb <==
## **Introduction to ML for NLP [Network + Practical]**

### **Linear Support Vector Classifier**

#### **Libraries**

We have seen in notebook 7.A that we can correctly train a LinearSVC in each language.

We now have to tune the model, and store the best results.

# general
import pandas as pd

from utility.models_sklearn import CustomLinearSVC

print("> Libraries Imported")

#### **Setup**

We only need to import the dataset

dataframe = pd.read_pickle("data/3_multi_eurlex_encoded.pkl")
dataframe.head(3)

#### **Create a Grid Search for Fine Tuning**

First, we setup the needed parameters.

LANGUAGE_LIST = ["en", "de", "it", "pl", "sv"]
C_LIST = [0.1, 1, 5, 10, 50, 100, 1000]

Then, we execute a grid search.

# setup placeholders for results
df_language_list = []
df_c_list = []
df_mean_acc = []

# start grid search
for LANGUAGE in LANGUAGE_LIST:
    for C in C_LIST:

        # instantiate the model
        LinearSVC_temp = CustomLinearSVC(
            dataset     = dataframe,
            language    = LANGUAGE,

            C           = C # only hyperparameter
        )

        # train the model
        eval_res = LinearSVC_temp.train_model()

        # save results
        df_language_list.append(LANGUAGE)
        df_c_list.append(C)
        df_mean_acc.append(eval_res["macro avg"]["precision"])

# save results in a dataframe
global_res_df = pd.DataFrame(
    list(zip(df_language_list,df_c_list,df_mean_acc)),
    columns = ["language", "c", "validation_accuracy (mean)"]
    )

# save them and show them
global_res_df.to_csv("models/LinearSVC/training_results.csv")
global_res_df
