==> 02_network_and_practical/src/07.C_LinearSVC_explore_best_model.ipynb <==
## **Introduction to ML for NLP [Network + Practical]**

### **Linear Support Vector Classifier**

#### **Libraries**

After the fine-tuning phase, we know what the best C values are for each languages.

In this notebook, we test the best models on a test set they have never seen, in order to verify their real performance.

# general
import pandas as pd

from utility.models_sklearn import CustomLinearSVC

print("> Libraries Imported")

#### **Setup**

We only need to import the dataset

dataframe = pd.read_pickle("data/3_multi_eurlex_encoded.pkl")
dataframe.head(3)

#### **Test Best Models**

##### *English Model*

LinearSVC_EN = CustomLinearSVC(
    dataset     = dataframe,
    language    = "en",
    C           = 1.0 # only hyperparameter
)

eval_res = LinearSVC_EN.train_model()
test_res = LinearSVC_EN.test_model(on="test")

##### *German Model*

LinearSVC_DE = CustomLinearSVC(
    dataset     = dataframe,
    language    = "de",
    C           = 1.0 # only hyperparameter
)

eval_res = LinearSVC_DE.train_model()
test_res = LinearSVC_DE.test_model(on="test")

##### *Italian Model*

LinearSVC_IT = CustomLinearSVC(
    dataset     = dataframe,
    language    = "it",
    C           = 1.0 # only hyperparameter
)

eval_res = LinearSVC_IT.train_model()
test_res = LinearSVC_IT.test_model(on="test")

##### *Polish Model*

LinearSVC_PL = CustomLinearSVC(
    dataset     = dataframe,
    language    = "pl",
    C           = 1.0 # only hyperparameter
)

eval_res = LinearSVC_PL.train_model()
test_res = LinearSVC_PL.test_model(on="test")

##### *Swedish Model*

LinearSVC_SV = CustomLinearSVC(
    dataset     = dataframe,
    language    = "sv",
    C           = 1.0 # only hyperparameter
)

eval_res = LinearSVC_SV.train_model()
test_res = LinearSVC_SV.test_model(on="test")
